git and index support (#2)
Add support for an index file: the index file contain associations between directories and remote repositories (at given references). The directories are automatically made relative the directory containing the index file. When using an index file, target directories are automatically made read-only, as `ungit` is then used like a simplified package manager.

When the target directory is located within a `git` repository, the index file mode is automatically turned on and will be a file called `.ungit` at the root of the `git` repository.

To better fit the more complex behaviour, `ungit` now takes a command as a first argument. Commands can be:
+ add: as described above.
+ delete (or remove): remove a directory, remove it from index.
+ install: install all associations from a given index, if missing.